Add Allwinner H5 support
Allwinner H5 is a SoC like H3, but with Cortex-A7 cores replaced with
Cortex-A53 cores, and a memory map change.

Add support for it, it can use the same UART pinmux with H3.

The address of SRAM A1 on H5 seems to be not 0x40000, it needs more
investigation. Currently I disabled SRAM A1 test on H5.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>